drm/rockchip: drv: Convert the order of colorspace information read in edid
The definitions of BT2020_RGB and BT2020_YCC in DRM are in the
opposite order to that in EDID. so the values of two bits need to
be exchanged.

+++ drivers/gpu/drm/rockchip/dw_hdmi-rockchip.c
@@ -372,7 +372,7 @@ struct rockchip_hdmi {
 	struct rockchip_drm_sub_dev sub_dev;
 
 	u64 force_frl_rate;
-	u8 edid_colorimetry;
+	u32 edid_colorimetry;
 	u8 hdcp_status;
 	u8 dovi_vsdb[DOVI_VSDB_LEN];
 	struct hdr10_plus_vsdb hdr10_plus_data;
