+++ drivers/gpu/drm/rockchip/rockchip_drm_drv.h
@@ -757,7 +757,7 @@ int rockchip_drm_get_yuv422_format(struct drm_connector *connector,
 int rockchip_drm_parse_cea_ext(struct rockchip_drm_hdmi21_data *hdmi21_data,
 			       const struct edid *edid);
 int rockchip_drm_parse_dovi(u8 *sink_data, const struct edid *edid);
-int rockchip_drm_parse_colorimetry_data_block(u8 *colorimetry, const struct edid *edid);
+int rockchip_drm_parse_colorimetry_data_block(u32 *colorimetry, const struct edid *edid);
 u8 rockchip_drm_parse_hdr10_plus_vsdb(const struct edid *edid);
 struct dma_buf *rockchip_drm_gem_prime_export(struct drm_gem_object *obj, int flags);
 long rockchip_drm_dclk_round_rate(u32 version, struct clk *dclk, unsigned long rate);
